# specs/001-task-management-api/contracts/task-management.openapi.yaml
openapi: 3.1.0
info:
  title: VistaFlow Task Management API
  version: 1.0.0
paths:
  /task-types:
    get:
      summary: List supported task types
      responses:
        '200':
          description: Task type catalog
  /tasks:
    get:
      summary: List task definitions
      responses:
        '200':
          description: Task definition list
    post:
      summary: Create task definition
      responses:
        '201':
          description: Task created
        '409':
          description: Duplicate task name
  /tasks/{taskId}:
    get:
      summary: Get task definition detail
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task detail
        '404':
          description: Task not found
    patch:
      summary: Update task definition
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task updated
        '404':
          description: Task not found
        '409':
          description: Duplicate task name or task currently running with conflicting change
    delete:
      summary: Delete task definition
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted
        '404':
          description: Task not found
        '409':
          description: Task currently running
  /tasks/{taskId}/run:
    post:
      summary: Trigger a manual task run
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      responses:
        '202':
          description: Run accepted
        '404':
          description: Task not found
        '409':
          description: Task disabled or already running
  /tasks/{taskId}/runs:
    get:
      summary: List runs for a task
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task run list
        '404':
          description: Task not found
  /task-runs/{runId}:
    get:
      summary: Get a single task run
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task run detail
        '404':
          description: Run not found
  /task-runs/{runId}/logs:
    get:
      summary: List logs for a task run
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Run log list
        '404':
          description: Run not found
  /task-runs/{runId}/terminate:
    post:
      summary: Terminate a running task run
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: integer
      responses:
        '202':
          description: Termination accepted
        '404':
          description: Run not found
        '409':
          description: Run not active or cannot be terminated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []